Send mapping records for all EIDs in Map-Replies
Previously, a Map-Reply with an IPv6 EID mapping reccord would have been
sent, of an IPv6 EID was configured, an IPv4 record otherwise. This
commit changes the behaviour, so that one EID for each AFI is returned,
when available.

Code should be changed so that reply depends on the actual request.