Create a basic Vagrant setup for development
This commit adds a `Vagrantfile` to the top-level directory of the tree
to allow quick creation of a VM with all prerequisites preinstalled so
the code can be built quickly in an isolated environment. This is
especially useful for development on platforms other than Linux, where
code editing is done on the native host OS, and the Linux VM created by
Vagrant is used only for building and testing the code.

Vagrant automcatically sets up a shared folder `/vagrant` pointing to
the folder where the Vagrantfile resides, in this case the top-level
folder of the source tree. This way, changes to the source code on the
host computer are automatically reflected in the VM, and binaries
created in the VM are also automatically available to the host OS. This
can be especially helpful for Android development on non-Linux
platforms.

Signed-off-by: Lorand Jakab <lojakab@cisco.com>