Everything should be "working" except for the recieve_map_referral. By "working" I mean that I can't really say if the addresses where the messages are being sent, or the addresses listed as source/dest are completely correct, we'll look at that later. There are some "garbage" OOR_LOG left over, I think I'll keep them at least until the next revision, as they could prove useful until then.
One issue I have found is that when deleting an item(for example, a "ddt_pending_request" it doesn't automatically get removed from the Map Resolver's pending request database. When doing "free", the database entry still exists, and is pointing to the freed space. When trying to get that database entry, as it only needs to return a "void" type, it pulls whatever bits are there(which are not considered NULL), and gives that. As a result, it appears as there is an entry, even though its contents had been removed. I encountered this issue when sending a map request for an eid after the previous pending request had already been "deleted", the MR would think a Pending Request for that EID still exists in the database, and try to add an Original Request to it from the newly recieved message. The entry of the database, being pulled from a formerly freed space, obviously didn't have the "original_requests" member, so it crashed.  For all this, I have added a pointer to the Map Resolver object in the "del" functions of both the Pending Request and the Mref Cache Entry, so they can remove themselves from their respective db before freeing their space.
If the "del" function has to STRICTLY have only the item to be deleted as a parameter, then I'll have no choice but to add a pointer to the Map Resolver as a member of both the Pending Request and the Mref Cache Entry, so I may access its databases from the del function.