Add Vagrant setup for a quick testbed
This commit adds a Vagrant setup to create a quick OOR testbed with
three VMs with OOR running, pre-configured for various roles: one map
server, and two mobile nodes. The VMs are also provisioned with a
.bash_history file, so one can just SSH in, press "arrow up" a few
times and ping one mobile node from the other, or monitor traffic with
TShark easily.

Signed-off-by: Lorand Jakab <lojakab@cisco.com>