Vagrant: harmonize /vagrant synced folder
By default the /vagrant folder inside the VM is pointed to the directory
of the Vagrantfile used to bring up the VM. As a result the VMs created
by the two different Vagrantfiles we have point to different locations.
Thanks makes sharing some of the provisioning scripts harder.

This patch changes the configuration so that all VMs point /vagrant to
the root of the git tree, and updates the provisioning scripts
accordingly.

Signed-off-by: Lorand Jakab <lojakab@cisco.com>