Vagrant: Add Android emulator node to testbed
This is only initial work to get things going, so far it is poossible to
bring up a VM with the Android SDK installed, generate an Android
Virtual Device (AVD) and start the emulator in headless mode.

For some reason, installing the OOR .apk doesn't work, it may be because
the emulator is very slow, since it doesn't work in hardware accelerated
mode, even with x86 system images. The reason behind the lack of
hardware acceleration support is that VirtualBox 5.2.x doesn't expose
the VMX CPU extension in the guest Linux OS (`grep vmx /proc/cpuinfo`).

Another issue for future work is figuring out how to connect the
emulator network stack to the rest of the LISP testedbed in a good way,
since the emulator is behind a virtual router, doing NAT:

https://developer.android.com/studio/run/emulator-networking.html

Signed-off-by: Lorand Jakab <lojakab@cisco.com>