Vagrant: make IPv6 optional, default off
With certain version/OS combinations bringing up the Vagrant testbed
fails with a message like:

/usr/lib64/ruby/2.3.0/resolv-replace.rb:42:in `connect': Invalid argument - connect(2) for "fe80::ffff:ffff:ffff:ffff" port 80 (Errno::EINVAL)

This patch introduces a boolean variable in the Vagrantfile to turn IPv6
setup off by default. Users who want to work with IPv6 can enable it,
and try to see if it works in their environment, but others don't have
to try to debug cryptic error messages.

Signed-off-by: Lorand Jakab <lojakab@cisco.com>