Use correct source port and encap for Data-Map-Registers
The source port of the outer header of Data-Map-Register messages was
hardcoded to the LISP data port number, and the encapsulation was always
LISP, so NAT-Traversal would not work when LISP-GPE was used for data
encapsulation.

This patch adds support for NAT-Traversal when LISP-GPE encapsulation is
used, but requires an RTR with support for VXLAN-GPE, in addition to the
xTRs/MNs.

The current encapsulation port number is saved in the lisp_tr structure,
in addition to the existing encapsulation type on startup, and is used
as source port when building Data-Map-Register messages, so that the
correct NAT state is established.

When re-encapsulating Data-Map-Notify messages, it now uses the correct
data encapsulation towards the xTR/MN.

Signed-off-by: Lorand Jakab <lojakab@cisco.com>